+++ src/main/webapp/WEB-INF/views/list-orders.jsp
@@ -6,16 +6,27 @@
 <head>
 <meta charset="ISO-8859-1">
 <title>Orders List</title>
+<link rel = "stylesheet"
+         href = "https://maxcdn.bootstrapcdn.com/bootstrap/4.1.1/css/bootstrap.min.css">
+      <script src = "https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
+      <script src = "https://maxcdn.bootstrapcdn.com/bootstrap/4.1.1/js/bootstrap.min.js"></script>
 <style>
 <%@include file="/WEB-INF/css/listyarn.css"%>
+ body {
+    background-image:
+        url(https://st.depositphotos.com/10614052/54679/i/450/depositphotos_546796424-stock-photo-thread-spools-light-background.jpg);
+    background-size: cover; 
+    background-repeat:no-repeat; 
+    
+   }  
 </style>
 <link rel="stylesheet"
 	href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
 </head>
 <body>
-	<div class=top >
+	<div style="width:1349px"class=top >
 		<div>Orders List</div>
-		<div><a href="/yarn/yarnindex"><em class="fa fa-home"
+		<div><a href="/yarn/mypreviousyarnindex"><em class="fa fa-home"
 		style="font-size: 30px;float:right; color: white;"></em></a></div>
 	</div>
 	<br>
@@ -35,9 +46,9 @@
 					<th>Receiving Date</th>
 					<th>Order Status</th>
 					<th>Advance</th>
+					<th>Invoice</th>
 					<th>Edit</th>
 					<th>Delete</th>
-
 				</tr>
 			</thead>
 			<tbody>
@@ -47,15 +58,16 @@
 						<td>${orders.customerId}</td>
 						<td>${orders.orderId}</td>
 						<td>${orders.orderDate}</td>
-						<td>${orders.orderedCustomer}</td>
+						<td>${orders.name}</td>
 						<td>${orders.orderedQuantity}</td>
 						<td>${orders.rate}</td>
 						<td>${orders.orderAmount}</td>
 						<td>${orders.receivingDate}</td>
 						<td>${orders.orderStatus}</td>
 						<td>${orders.advance}</td>
-						<td><a class=text href="updateform?id=${orders.orderId}">Edit</a></td>
-						<td><a class=text href="deleteorder?id=${orders.orderId}">Delete</a></td>
+						<td><a class=text href="/invoice/getordersinvoice?id=${orders.orderId}"><button type="button" class="btn btn-primary">Invoice</button></a></td>
+						<td><a class=text href="updateform?id=${orders.orderId}"><button type="button" class="btn btn-dark">Edit</button></a></td>
+						<td><a class=text href="deleteorder?id=${orders.orderId}"onclick="return confirm('Are you sure you want to delete order ${orders.orderId} ?');"><button type="button" class="btn btn-danger">Delete</button></a></td>
 					</tr>
 				</c:forEach>
 			</tbody>
