+++ src/main/webapp/WEB-INF/views/list-suppliers.jsp
@@ -6,16 +6,27 @@
 <head>
 <meta charset="ISO-8859-1">
 <title>Supplier List</title>
+<link rel = "stylesheet"
+         href = "https://maxcdn.bootstrapcdn.com/bootstrap/4.1.1/css/bootstrap.min.css">
+      <script src = "https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
+      <script src = "https://maxcdn.bootstrapcdn.com/bootstrap/4.1.1/js/bootstrap.min.js"></script>
 <style>
 <%@include file="/WEB-INF/css/listyarn.css"%>
+ body {
+    background-image:
+        url(https://st.depositphotos.com/10614052/54679/i/450/depositphotos_546796424-stock-photo-thread-spools-light-background.jpg);
+    background-size: cover; 
+    background-repeat:no-repeat; 
+    
+   }  
 </style>
 <link rel="stylesheet"
 	href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
 </head>
 <body>
-	<div class=top >
+	<div style="width:1349px"class=top >
 		<div>Suppliers List</div>
-		<div><a href="/yarn/yarnindex"><em class="fa fa-home"
+		<div><a href="/yarn/mypreviousyarnindex"><em class="fa fa-home"
 		style="font-size: 30px;float:right; color: white;"></em></a></div>
 	</div>
 	<br>
@@ -37,6 +48,7 @@
 					<th>Bank Name</th>
 					<th>Account Number</th>
 					<th>IFSC Code</th>
+					<th>Edit</th>
 					<th>Delete</th>
 				</tr>
 			</thead>
@@ -56,8 +68,9 @@
 						<td>${supplier.bankName}</td>
 						<td>${supplier.accountNumber}</td>
 						<td>${supplier.ifscCode}</td>
+						<td><a class=text href="updateform?id=${supplier.supplierId}"><button type="button" class="btn btn-dark">Edit</button></a></td>
 						<td><a class=text
-							href="deletesupplier?id=${supplier.supplierId}">Delete</a></td>
+							href="deletesupplier?id=${supplier.supplierId}"onclick="return confirm('Are you sure you want to delete supplier ${supplier.supplierId} ?');"><button type="button" class="btn btn-danger">Delete</button></a></td>
 					</tr>
 				</c:forEach>
 			</tbody>
